requirements.txt
tweepy==4.14.0
anthropic==0.41.1
pillow==11.0.0
python-dotenv==1.0.0
